server/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
supabase==1.0.3
requests==2.31.0
reportlab==4.0.7
gunicorn==21.2.0
